+++ src/Spreads.Collections/Collections/SortedHashMap.Obsolete.fs
@@ -165,7 +165,7 @@ type SortedHashMap<'K,'V when 'K : comparison>
   override this.GetCursor() : ICursor<'K,'V> =  this.GetROOMCursor() :> ICursor<'K,'V>
 
   member private this.GetROOMCursor() : MapCursor<'K,'V> =
-    let outer = ref (outerMap.GetCursor())
+    let outer = ref (outerMap.GetCursor() :> ICursor<_,_>)
     outer.Value.MoveFirst() |> ignore // otherwise initial move is skipped in MoveAt, isReset knows that we haven't started in SHM even when outer is started
     let inner = ref Unchecked.defaultof<ICursor<'K, 'V>> // ref (outer.Value.CurrentValue.GetPointer())
     // TODO (perf) pointers must own previous idx/bucket, SHM methods must
@@ -189,7 +189,7 @@ type SortedHashMap<'K,'V when 'K : comparison>
               true
             else
               if outer.Value.MoveNext() then // go to the next bucket
-                inner := outer.Value.CurrentValue.GetCursor()
+                inner := outer.Value.CurrentValue.GetCursor():> ICursor<_,_>
                 let res = inner.Value.MoveFirst()
                 if res then
                   currentKey := inner.Value.CurrentKey
@@ -214,7 +214,7 @@ type SortedHashMap<'K,'V when 'K : comparison>
               true
             else
               if outer.Value.MovePrevious() then // go to the previous bucket
-                inner := outer.Value.CurrentValue.GetCursor()
+                inner := outer.Value.CurrentValue.GetCursor():> ICursor<_,_>
                 let res = inner.Value.MoveLast()
                 if res then
                   currentKey := inner.Value.CurrentKey
@@ -236,7 +236,7 @@ type SortedHashMap<'K,'V when 'K : comparison>
           let res =
             if c <> 0 || !isReset then // not in the current bucket, switch bucket
               if outer.Value.MoveAt(newHash, Lookup.EQ) then // Equal!
-                inner := outer.Value.CurrentValue.GetCursor()
+                inner := outer.Value.CurrentValue.GetCursor():> ICursor<_,_>
                 inner.Value.MoveAt(newSubIdx, direction)
               else
                 false
@@ -254,7 +254,7 @@ type SortedHashMap<'K,'V when 'K : comparison>
               | Lookup.LT | Lookup.LE ->
                 // look into previous bucket
                 if outer.Value.MovePrevious() then
-                  inner := outer.Value.CurrentValue.GetCursor()
+                  inner := outer.Value.CurrentValue.GetCursor():> ICursor<_,_>
                   let res = inner.Value.MoveAt(newSubIdx, direction)
                   if res then
                     currentKey := inner.Value.CurrentKey
@@ -269,7 +269,7 @@ type SortedHashMap<'K,'V when 'K : comparison>
               | Lookup.GT | Lookup.GE ->
                 // look into next bucket
                 if outer.Value.MoveNext() then
-                  inner := outer.Value.CurrentValue.GetCursor()
+                  inner := outer.Value.CurrentValue.GetCursor():> ICursor<_,_>
                   let res = inner.Value.MoveAt(newSubIdx, direction)
                   if res then
                     currentKey := inner.Value.CurrentKey